[ZH] Fix type mismatch errors for Zero Hour build (#426)

+++ GeneralsMD/Code/Libraries/Source/WWVegas/WW3D2/render2dsentence.cpp
@@ -1720,8 +1720,8 @@ FontCharsClass::Grow_Unicode_Array (WCHAR ch)
 		return ;
 	} 
 
-	uint16 first_index	= min( FirstUnicodeChar, ch );
-	uint16 last_index		= max( LastUnicodeChar, ch );
+	uint16 first_index	= min( FirstUnicodeChar, static_cast<uint16>(ch) );
+	uint16 last_index		= max( LastUnicodeChar, static_cast<uint16>(ch) );
 	uint16 count			= (last_index - first_index) + 1;
 
 	//
